server/requirements.txt
certifi==2019.11.28
click==7.1.2
dnspython==1.16.0
Flask==1.1.2
Flask-Cors==3.0.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pipenv==2018.11.26
pykerberos==1.2.1
pymongo==3.11.0
python-dotenv==0.14.0
six==1.15.0
virtualenv==16.7.9
virtualenv-clone==0.5.3
Werkzeug==1.0.1
